fix: api 컨트롤러 user -> auth 이동
