+++ src/main/java/shop/wannab/userservice/auth/controller/AuthController.java
@@ -6,6 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -98,6 +99,13 @@ public class AuthController {
     public ResponseEntity<TokenPayloadResponse> info(@RequestBody @Valid TokenPayloadRequest tokenPayloadRequest) {
         TokenPayloadResponse response = authService.getTokenPayload(tokenPayloadRequest);
         return ResponseEntity.ok(response);
+    }
+
+    @PutMapping("/lastLogin")
+    public ResponseEntity<Void> updateLastLogin(@RequestParam Long userId) {
+        log.info("Controller: updateLastLogin");
+        authService.updateLastLogin(userId);
+        return ResponseEntity.noContent().build();
     }
 
 }
