+++ src/main/java/shop/wannab/userservice/user/controller/UserController.java
@@ -8,7 +8,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -63,10 +62,4 @@ public class UserController {
         return userService.birthUserList(month);
     }
 
-    @PutMapping("/lastLogin")
-    public ResponseEntity<Void> updateLastLogin(@RequestHeader(HeaderUtil.HEADER_ID_NAME) Long userId) {
-        log.info("Controller: updateLastLogin");
-        userService.updateLastLogin(userId);
-        return ResponseEntity.noContent().build();
-    }
 }
